ADD: Controllers

+++ WebShop/Services/Implementations/CommentService.cs
@@ -1,8 +1,8 @@
 ﻿using AutoMapper;
 using Microsoft.AspNetCore.Identity;
-using Microsoft.EntityFrameworkCore.ChangeTracking;
 using WebShop.Dtos.Read;
 using WebShop.Dtos.Write;
+using WebShop.Exceptions;
 using WebShop.Models;
 using WebShop.Repositories.Interfaces;
 using WebShop.Services.Interfaces;
@@ -37,13 +37,19 @@ namespace WebShop.Services.Implementations
             comment.CreatedAt = DateTime.UtcNow;
             comment.UpdatedAt = DateTime.UtcNow;
 
-            return await _commentRepository.AddAsync( comment );
+            return await _commentRepository.AddAsync(comment);
         }
 
         public async Task<bool> DeleteAsync(int commentId)
         {
-            Comment comment = await _commentRepository.GetAsync( commentId ); 
-            return await _commentRepository.DeleteAsync( comment );
+            Comment comment = await _commentRepository.GetAsync(commentId);
+
+            if (comment == null)
+            {
+                throw new NotFoundException($"Отзыв {commentId} не найден");
+            }
+
+            return await _commentRepository.DeleteAsync(comment);
         }
 
         public async Task<List<CommentR>> GetAllAsync()
@@ -67,6 +73,10 @@ namespace WebShop.Services.Implementations
         public async Task<List<CommentR>> GetByFeedbackAsync(int feedbackId)
         {
             Feedback feedback = await _feedbackRepository.GetAsync(feedbackId);
+
+            if (feedback == null)
+                throw new NotFoundException($"Отзыв {feedbackId} не существует");
+
             List<Comment> comments = await _commentRepository.GetByFeedBackAsync(feedback);
 
             List<CommentR> commentDtos = new List<CommentR>();
@@ -80,6 +90,10 @@ namespace WebShop.Services.Implementations
         public async Task<List<CommentR>> GetByParentCommentAsync(int parentCommentId)
         {
             Comment parentComment = await _commentRepository.GetAsync(parentCommentId);
+
+            if (parentComment == null)
+                throw new NotFoundException($"Отзыв {parentCommentId} не существует");
+
             List<Comment> comments = await _commentRepository.GetByParentCommentAsync(parentComment);
 
             List<CommentR> commentDtos = new List<CommentR>();
@@ -93,6 +107,10 @@ namespace WebShop.Services.Implementations
         public async Task<List<CommentR>> GetByUserAsync(string userId)
         {
             User user = await _userManager.FindByIdAsync(userId);
+
+            if (user == null)
+                throw new NotFoundException($"Отзыв {userId} не существует");
+
             List<Comment> comments = await _commentRepository.GetByUserAsync(user);
 
             List<CommentR> commentDtos = new List<CommentR>();
@@ -106,7 +124,7 @@ namespace WebShop.Services.Implementations
         public async Task<bool> UpdateAsync(CommentW commentDto)
         {
             Comment comment = await MapFromDto(commentDto);
-            comment.UpdatedAt = DateTime.UtcNow; 
+            comment.UpdatedAt = DateTime.UtcNow;
 
             return await _commentRepository.UpdateAsync(comment);
         }
@@ -118,7 +136,7 @@ namespace WebShop.Services.Implementations
                 Text = commentDto.Text,
                 User = await _userManager.FindByIdAsync(commentDto.UserId),
                 Feedback = await _feedbackRepository.GetAsync(commentDto.FeedbackId),
-                ParentComment = commentDto.ParentCommentId == 0 ? null 
+                ParentComment = commentDto.ParentCommentId == 0 ? null
                         : await _commentRepository.GetAsync(commentDto.ParentCommentId),
                 Product = await _productRepository.GetAsync(commentDto.ProductId)
             };
@@ -128,7 +146,7 @@ namespace WebShop.Services.Implementations
         {
             CommentR commentR = _mapper.Map<CommentR>(comment);
             commentR.User = _mapper.Map<UserR>(comment.User);
-            return commentR;    
+            return commentR;
         }
     }
 }
